Edit: Moved contexts implementations to view layer
