+++ src/modules/chooseGame/infra/controllers/useFetchPlayersController.ts
@@ -1,7 +1,7 @@
 import makeFetchPlayersService from 'modules/chooseGame/services/factories/makeFetchPlayersService';
 import { useCallback, useEffect, useMemo, useState } from 'react';
-import useSessionContext from 'shared/container/contexts/SessionContext/contexts/useSessionContext';
-import useToastContext from 'shared/container/contexts/ToastContext/contexts/useToastContext';
+import { useSessionContext } from 'shared/view/contexts';
+import { useToastContext } from 'shared/view/contexts';
 import IPlayer from 'shared/domain/entities/IPlayer';
 
 interface UseFetchPlayersController {
