+++ src/modules/chooseGame/infra/controllers/useGetUserController.ts
@@ -1,7 +1,7 @@
 import makeGetUserService from 'modules/chooseGame/services/factories/makeGetUserService';
 import { useCallback, useEffect, useMemo, useState } from 'react';
-import useSessionContext from 'shared/container/contexts/SessionContext/contexts/useSessionContext';
-import useToastContext from 'shared/container/contexts/ToastContext/contexts/useToastContext';
+import { useSessionContext } from 'shared/view/contexts';
+import { useToastContext } from 'shared/view/contexts';
 import IUser from 'shared/domain/entities/IUser';
 
 interface UserParams {
