+++ src/modules/chooseGame/infra/controllers/useUpdateUserController.ts
@@ -1,7 +1,7 @@
 import makeUpdateUserService from 'modules/chooseGame/services/factories/makeUpdateUserService';
 import { useCallback, useMemo, useState } from 'react';
-import useSessionContext from 'shared/container/contexts/SessionContext/contexts/useSessionContext';
-import useToastContext from 'shared/container/contexts/ToastContext/contexts/useToastContext';
+import { useSessionContext } from 'shared/view/contexts';
+import { useToastContext } from 'shared/view/contexts';
 
 interface IUpdateValues {
   firstname: string;
