+++ src/modules/chooseGame/view/pages/Lobby/components/Footer/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 // Hooks
 import { useNavigation } from '@react-navigation/native';
-import useSessionContext from 'shared/container/contexts/SessionContext/contexts/useSessionContext';
+import { useSessionContext } from 'shared/view/contexts';
 
 // styles
 import { Row, Button } from './styles';
