+++ src/modules/selectedGame/infra/controllers/useGetFeedPostsController.ts
@@ -1,8 +1,9 @@
+import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+
 import IFeedPost from 'modules/selectedGame/entities/IFeedPost';
 import makeGetFeedPostsService from 'modules/selectedGame/services/factories/makeGetFeedPostsService';
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
-import useSessionContext from 'shared/container/contexts/SessionContext/contexts/useSessionContext';
-import useToastContext from 'shared/container/contexts/ToastContext/contexts/useToastContext';
+import { useSessionContext } from 'shared/view/contexts';
+import { useToastContext } from 'shared/view/contexts';
 
 interface UseGetFeedPostsController {
   loading: boolean;
