+++ src/modules/selectedGame/infra/controllers/useRequestAchievementUnlockController.ts
@@ -1,8 +1,9 @@
+import { useCallback, useMemo, useState } from 'react';
+
 import IRequestAchievementDTO from 'modules/selectedGame/dtos/IRequestAchievementDTO';
 import makeRequestAchievementUnlockService from 'modules/selectedGame/services/factories/makeRequestAchievementUnlockService';
-import { useCallback, useMemo, useState } from 'react';
-import useSessionContext from 'shared/container/contexts/SessionContext/contexts/useSessionContext';
-import useToastContext from 'shared/container/contexts/ToastContext/contexts/useToastContext';
+import { useSessionContext } from 'shared/view/contexts';
+import { useToastContext } from 'shared/view/contexts';
 
 interface UseRequestAchievementUnlockController {
   loading: boolean;
