+++ src/modules/selectedGame/view/pages/RequestActivity/index.tsx
@@ -1,21 +1,22 @@
 import React, { useCallback } from 'react';
 
 // Entities
+import { useRoute, RouteProp, useNavigation } from '@react-navigation/native';
+import { Formik } from 'formik';
+
 import IActivity from 'modules/selectedGame/entities/IActivity';
 
 // Components
-import { Formik } from 'formik';
+import useRequestActivityController from 'modules/selectedGame/infra/controllers/useRequestActivityController';
+import RequestActivitySchema from 'modules/selectedGame/view/validation/RequestActivitySchema';
 import { Input } from 'shared/view/components';
+import { useSessionContext, useToastContext } from 'shared/view/contexts';
+
 import { DateInput, Footer } from './components';
 
 // Hooks
-import { useRoute, RouteProp, useNavigation } from '@react-navigation/native';
-import useRequestActivityController from 'modules/selectedGame/infra/controllers/useRequestActivityController';
-import useToastContext from 'shared/container/contexts/ToastContext/contexts/useToastContext';
-import useSessionContext from 'shared/container/contexts/SessionContext/contexts/useSessionContext';
 
 // Schemas
-import RequestActivitySchema from 'modules/selectedGame/view/validation/RequestActivitySchema';
 
 // Styles
 import { Container, Title, Paragraph, Info, Form } from './styles';
