+++ src/shared/view/contexts/SessionContext/hooks/useSessionContext.ts
