+++ src/shared/view/contexts/SessionContext/implementations/DefaultSessionContext/index.tsx
@@ -1,11 +1,11 @@
 import React, { useCallback, useEffect, useMemo, useState } from 'react';
 
-import { SessionContextProvider } from 'shared/container/contexts/SessionContext/contexts/useSessionContext';
-import useThemeContext from 'shared/container/contexts/ThemeContext/contexts/useThemeContext';
 import ISessionContext from 'shared/domain/providers/ISessionContext';
 import makeHttpProvider from 'shared/infra/providers/factories/makeHttpProvider';
 import makeJwtProvider from 'shared/infra/providers/factories/makeJwtProvider';
 import makeStorageProvider from 'shared/infra/providers/factories/makeStorageProvider';
+import { useThemeContext } from 'shared/view/contexts';
+import { SessionContextProvider } from 'shared/view/contexts/SessionContext/hooks/useSessionContext';
 
 const USER_STORAGE_KEY = '@GameTask/token';
 const GAME_STORAGE_KEY = '@GameTask/game';
