+++ src/shared/view/contexts/SessionContext/index.ts
@@ -1,0 +1,4 @@
+import useSessionContext from 'shared/view/contexts/SessionContext/hooks/useSessionContext';
+import SessionContext from 'shared/view/contexts/SessionContext/implementations/DefaultSessionContext';
+
+export { SessionContext, useSessionContext };
