+++ src/shared/view/contexts/ThemeContext/implementations/StyledComponentsThemeContext/helpers/getNewPalette.ts
