+++ src/shared/view/contexts/ThemeContext/implementations/StyledComponentsThemeContext/helpers/index.ts
