+++ src/shared/view/contexts/ThemeContext/index.ts
@@ -1,0 +1,4 @@
+import useThemeContext from 'shared/view/contexts/ThemeContext/hooks/useThemeContext';
+import ThemeContext from 'shared/view/contexts/ThemeContext/implementations/StyledComponentsThemeContext';
+
+export { ThemeContext, useThemeContext };
