+++ src/shared/view/contexts/ToastContext/index.ts
@@ -1,0 +1,4 @@
+import useToastContext from 'shared/view/contexts/ToastContext/hooks/useToastContext';
+import ToastContext from 'shared/view/contexts/ToastContext/implementations/FlashMessageToastContext';
+
+export { ToastContext, useToastContext };
